linux-beagleboard_4.19: bump version 4.19.25 -> 4.19.31
Building lttng failed with:

	.../tmp-glibc/work/pocketbeagle-oe-linux-gnueabi/lttng-modules/2.10.9-r0/lttng-modules-2.10.9/probes/../blacklist/kprobes.h:19:4: error: #error "Your kernel is known to have buggy optimized kprobes implementation. Fixed by commit 0ac569bf6a7983c0c5747d6df8db9dc05bc92b6c \"ARM: 8834/1: Fix: kprobes: optimized kprobes illegal instruction\" in Linux. Disable CONFIG_OPTPROBES or upgrade your kernel."

In the lttng-modules sources, blacklist/kprobes.h, we find:

	#if LTTNG_KERNEL_RANGE(4,20,0, 4,20,13) \
		|| LTTNG_KERNEL_RANGE(4,19,9, 4,19,26) \
		|| LTTNG_KERNEL_RANGE(4,14,87, 4,14,104)
	# if defined(CONFIG_ARM) && defined(CONFIG_OPTPROBES)
	#  error "Your kernel is known to have buggy optimized kprobes implementation. Fixed by commit 0ac569bf6a7983c0c5747d6df8db9dc05bc92b6c \"ARM: 8834/1: Fix: kprobes: optimized kprobes illegal instruction\" in Linux. Disable CONFIG_OPTPROBES or upgrade your kernel."
	# endif /* #if defined(CONFIG_ARM) && defined(CONFIG_OPTPROBES) */
	#endif

Therefore I'm bumping the kernel version since 4.19.25 falls in this bad
range and 4.19.31 doesn't.

The beagleboard kernel has a lot of patches on top of the upstream 4.19.31,
one of which makes quite a mess of the device tree files. Therefore I've
reverted that patch from this series.

This build was tested on an actual pocketbeagle (Rev A2). The boot arrives at
a cmdline through the console successfully.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>